# back/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nareshka",
    "description": "Nareshka - платформа для изучения программирования",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/interview-categories/": {
      "get": {
        "tags": [
          "interview-categories"
        ],
        "summary": "Получить список категорий",
        "description": "Возвращает все категории вопросов интервью с статистикой",
        "operationId": "get_categories_api_v2_interview_categories__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  },
                  "type": "array",
                  "title": "Response Get Categories Api V2 Interview Categories  Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v2/interview-categories/statistics": {
      "get": {
        "tags": [
          "interview-categories"
        ],
        "summary": "Получить статистику категоризации",
        "description": "Возвращает общую статистику по категоризации вопросов",
        "operationId": "get_statistics_api_v2_interview_categories_statistics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesStatisticsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v2/interview-categories/cluster/{cluster_id}/questions": {
      "get": {
        "tags": [
          "interview-categories"
        ],
        "summary": "Получить вопросы кластера",
        "description": "Возвращает список вопросов определенного кластера",
        "operationId": "get_cluster_questions_api_v2_interview_categories_cluster__cluster_id__questions_get",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID кластера",
              "title": "Cluster Id"
            },
            "description": "ID кластера"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы",
              "default": 1,
              "title": "Page"
            },
            "description": "Номер страницы"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Количество вопросов на странице",
              "default": 50,
              "title": "Limit"
            },
            "description": "Количество вопросов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponse"
                  },
                  "title": "Response Get Cluster Questions Api V2 Interview Categories Cluster  Cluster Id  Questions Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/interview-categories/search/questions": {
      "get": {
        "tags": [
          "interview-categories"
        ],
        "summary": "Поиск вопросов",
        "description": "Полнотекстовый поиск по вопросам интервью с фильтрацией и пагинацией",
        "operationId": "search_questions_api_v2_interview_categories_search_questions_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Поисковый запрос (минимум 1 символ, * для всех)",
              "title": "Q"
            },
            "description": "Поисковый запрос (минимум 1 символ, * для всех)"
          },
          {
            "name": "category_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по ID категорий (множественный)",
              "title": "Category Ids"
            },
            "description": "Фильтр по ID категорий (множественный)"
          },
          {
            "name": "cluster_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по ID кластеров (множественный)",
              "title": "Cluster Ids"
            },
            "description": "Фильтр по ID кластеров (множественный)"
          },
          {
            "name": "companies",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по названиям компаний (множественный)",
              "title": "Companies"
            },
            "description": "Фильтр по названиям компаний (множественный)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Максимальное количество результатов",
              "default": 50,
              "title": "Limit"
            },
            "description": "Максимальное количество результатов"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение для пагинации",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение для пагинации"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionsListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/interview-categories/companies/count": {
      "get": {
        "tags": [
          "interview-categories"
        ],
        "summary": "Получить общее количество компаний",
        "description": "Возвращает общее количество уникальных компаний в базе данных",
        "operationId": "get_total_companies_count_api_v2_interview_categories_companies_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Get Total Companies Count Api V2 Interview Categories Companies Count Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v2/interview-categories/clusters/all": {
      "get": {
        "tags": [
          "interview-categories"
        ],
        "summary": "Получить все кластеры",
        "description": "Возвращает список всех кластеров с возможностью фильтрации по категории",
        "operationId": "get_all_clusters_api_v2_interview_categories_clusters_all_get",
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по ID категории",
              "title": "Category Id"
            },
            "description": "Фильтр по ID категории"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по названию кластера",
              "title": "Search"
            },
            "description": "Поиск по названию кластера"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Максимальное количество кластеров",
              "default": 100,
              "title": "Limit"
            },
            "description": "Максимальное количество кластеров"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClusterResponse"
                  },
                  "title": "Response Get All Clusters Api V2 Interview Categories Clusters All Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/interview-categories/{category_id}": {
      "get": {
        "tags": [
          "interview-categories"
        ],
        "summary": "Получить детали категории",
        "description": "Возвращает подробную информацию о категории с кластерами и примерами",
        "operationId": "get_category_detail_api_v2_interview_categories__category_id__get",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID категории (например: javascript_core)",
              "title": "Category Id"
            },
            "description": "ID категории (например: javascript_core)"
          },
          {
            "name": "limit_questions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Количество примеров вопросов",
              "default": 10,
              "title": "Limit Questions"
            },
            "description": "Количество примеров вопросов"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/progress/categories": {
      "put": {
        "tags": [
          "Progress"
        ],
        "summary": "Обновление прогресса по категории",
        "description": "Обновление прогресса изучения категории",
        "operationId": "update_category_progress_api_v2_progress_categories_put",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryProgressUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Progress"
        ],
        "summary": "Получение прогресса по категориям",
        "description": "Получение прогресса изучения категорий",
        "operationId": "get_user_category_progress_api_v2_progress_categories_get",
        "parameters": [
          {
            "name": "main_category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по основной категории",
              "title": "Main Category"
            },
            "description": "Фильтр по основной категории"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryProgressResponse"
                  },
                  "title": "Response Get User Category Progress Api V2 Progress Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoriesStatisticsResponse": {
        "properties": {
          "total_questions": {
            "type": "integer",
            "title": "Total Questions",
            "description": "Всего вопросов"
          },
          "categorized_questions": {
            "type": "integer",
            "title": "Categorized Questions",
            "description": "Категоризировано вопросов"
          },
          "total_categories": {
            "type": "integer",
            "title": "Total Categories",
            "description": "Всего категорий"
          },
          "total_clusters": {
            "type": "integer",
            "title": "Total Clusters",
            "description": "Всего кластеров"
          },
          "categorization_rate": {
            "type": "number",
            "title": "Categorization Rate",
            "description": "Процент категоризации"
          }
        },
        "type": "object",
        "required": [
          "total_questions",
          "categorized_questions",
          "total_categories",
          "total_clusters",
          "categorization_rate"
        ],
        "title": "CategoriesStatisticsResponse",
        "description": "Статистика по категориям",
        "example": {
          "categorization_rate": 77.9,
          "categorized_questions": 6649,
          "total_categories": 13,
          "total_clusters": 182,
          "total_questions": 8532
        }
      },
      "CategoryDetailResponse": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/CategoryResponse",
            "description": "Информация о категории"
          },
          "clusters": {
            "items": {
              "$ref": "#/components/schemas/ClusterResponse"
            },
            "type": "array",
            "title": "Clusters",
            "description": "Кластеры категории"
          },
          "sample_questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Sample Questions",
            "description": "Примеры вопросов"
          }
        },
        "type": "object",
        "required": [
          "category",
          "clusters",
          "sample_questions"
        ],
        "title": "CategoryDetailResponse",
        "description": "Детальная информация о категории",
        "example": {
          "category": {
            "clusters_count": 60,
            "id": "javascript_core",
            "name": "JavaScript Core",
            "percentage": 24.4,
            "questions_count": 2082
          },
          "clusters": [
            {
              "category_id": "javascript_core",
              "id": 25,
              "keywords": [
                "closure"
              ],
              "name": "Замыкания",
              "questions_count": 38
            }
          ],
          "sample_questions": [
            {
              "company": "Яндекс",
              "id": "q123",
              "question_text": "Что такое замыкание?"
            }
          ]
        }
      },
...
      "CategoryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ID категории"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Название категории"
          },
          "questions_count": {
            "type": "integer",
            "title": "Questions Count",
            "description": "Количество вопросов"
          },
          "clusters_count": {
            "type": "integer",
            "title": "Clusters Count",
            "description": "Количество кластеров"
          },
          "percentage": {
            "type": "number",
            "title": "Percentage",
            "description": "Процент от общего числа вопросов"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color",
            "description": "Цвет категории для UI"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon",
            "description": "Иконка категории"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "questions_count",
          "clusters_count",
          "percentage"
        ],
        "title": "CategoryResponse",
        "description": "Модель ответа для категории",
        "example": {
          "clusters_count": 60,
          "color": "#f7df1e",
          "icon": "javascript",
          "id": "javascript_core",
          "name": "JavaScript Core",
          "percentage": 24.4,
          "questions_count": 2082
        }
      },
...
      "ClusterResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "ID кластера"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Название кластера"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id",
            "description": "ID категории"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords",
            "description": "Ключевые слова"
          },
          "questions_count": {
            "type": "integer",
            "title": "Questions Count",
            "description": "Количество вопросов"
          },
          "example_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Question",
            "description": "Пример вопроса"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category_id",
          "keywords",
          "questions_count"
        ],
        "title": "ClusterResponse",
        "description": "Модель ответа для кластера",
        "example": {
          "category_id": "javascript_core",
          "example_question": "Что такое замыкание в JavaScript?",
          "id": 25,
          "keywords": [
            "closure",
            "scope",
            "замыкание",
            "область",
            "видимости"
          ],
          "name": "Замыкания и область видимости",
          "questions_count": 38
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ID вопроса"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text",
            "description": "Текст вопроса"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company",
            "description": "Компания"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id",
            "description": "ID кластера"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id",
            "description": "ID категории"
          },
          "topic_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Name",
            "description": "Название топика"
          },
          "canonical_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Question",
            "description": "Канонический вопрос"
          },
          "interview_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Id",
            "description": "ID интервью"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_text"
        ],
        "title": "QuestionResponse",
        "description": "Модель ответа для вопроса",
        "example": {
          "canonical_question": "Что такое замыкание?",
          "category_id": "javascript_core",
          "cluster_id": 25,
          "company": "Яндекс",
          "id": "q123",
          "interview_id": "interview_Яндекс_001",
          "question_text": "Объясните, что такое замыкание в JavaScript",
          "topic_name": "Замыкания и область видимости"
        }
      },
      "QuestionsListResponse": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Questions",
            "description": "Список вопросов"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Общее количество вопросов"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Текущая страница"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Количество на странице"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "description": "Есть ли следующая страница"
          }
        },
        "type": "object",
        "required": [
          "questions",
          "total",
          "page",
          "limit",
          "has_next"
        ],
        "title": "QuestionsListResponse",
        "description": "Список вопросов с пагинацией",
        "example": {
          "has_next": true,
          "limit": 50,
          "page": 1,
          "questions": [],
          "total": 100
        }
      },
...
    }
  }
}